vagrant/rpm build - RPM builds on Fedora 19 VM are working
The builds should also work on the Fedora 17, 18, and 20 VMs that are
already built and hosted. Currently however the RPMs build do not
include a useful e.g. "fc19" suffix; hopefully I'll fix this in a
future commit.

Current state of play is that after installing the built RPM you can
run "vocalsalad" on a machine and it launches the Tornado Hello World
example bound to all interfaces port 20080.

Thie RPM is a dumb collection of a virtualenv that gets deployed to
/opt/vocalsalad, and then a softlink to the setup.py-generated
/opt/vocalsalad/bin/vocalsalad console script is created from
/usr/local/bin/vocalsalad. Using a virtualenv isolates us from
different versions of modules already installed but at the same time
we still require a working version of Python installed on the system.