systemd/rpm - provision systemd service, do proper upgrade
Provision a systemd service that starts the vocalsalad consolescript,
the main function in server.py. Moreover get the RPM to detect if it
is currently in an install/uninstall or an upgrade situation.
Upgrading.

-	If we're just installing from scratch copy the files, create the
	service, and start it.
- 	If we're just uninstalling from scratch delete the files,
	stop/disable the service.
-	If we're upgrading then we need to stop the service, copy the
	files, then restart the service. Very naughty if we just left the
	service running and copied/deleted files under its feet! And
	even worse if we forgot to restart the service.
-	If we're upgrading then the post-install step must not delete
	any files (!!).

Note that when installing RPMs reading [1] is a good tutorial on the
steps taking during upgrades/installs.

[1] https://www.ibm.com/developerworks/library/l-rpm2/